funzione proiezione ortogonale funzionante

+++ Vettore.h
@@ -23,8 +23,8 @@ void sottrarre_vettore(vettore *vetPtr, vettore *vetSub, int show);
 float prodotto_scalare_algebrico(vettore *vetPtrv, vettore *vetPtrw, int show);
 float prodotto_scalare_geometrico(vettore *vetPtrv, vettore *vetPtrw, int theta, int show);
 void normalizzare_vettore(vettore *vetPtr, int show);
-int magnitude_vettore(vettore *vetPtr, int show);
-double ottenere_angolo(vettore *vetPtrv, vettore *vetPtrw, int show);
+float magnitude_vettore(vettore *vetPtr, int show);
+float ottenere_angolo(vettore *vetPtrv, vettore *vetPtrw, int show);
 
 
 #endif
